GetByFilter can support author's name
